+++ src/Classes/Hourly.java
@@ -34,6 +34,7 @@ public class Hourly extends Employee {
   {
     timeCards.add(timecard);
     timeCards.peekLast().setWorked_week_time(timeCards);
+    getWorked_time();
   }
 
   public void paymentDefault()
