--- src/Functionalities/HelpFunctions.java
@@ -1,10 +1,0 @@
-package Functionalities;
-
-public class HelpFunctions {
-
-  public static void clear() {
-    for (int i = 0; i < 20; i++) {
-        System.out.println('\n');
-    }
-  }
-}
